Refactor: appDataS3Client -> S3Client 변경
